distsqlrun: add disk usage to windower stats
Adds maximum allocated bytes on disk metric to stats collected
during the windower run. Also, a minor refactor of instantiating
disk monitors in tests.

Release note: None
