+++ pkg/sql/logictest/testdata/logic_test/explain_analyze_plans
@@ -89,7 +89,7 @@ true
 query T
 SELECT url FROM [EXPLAIN ANALYZE (DISTSQL) SELECT avg(k) OVER () FROM kv]
 ----
-https://cockroachdb.github.io/distsqlplan/decode.html#eJy0k79O8zAUxffvKaw7fUiO8p_BU2FBFRJFBcGAMrjxVRWR2JHt0FZV3h3FGUpKC0FKR9s55_zOVe4epBL4wCs0wN4gBAoRUIiBQgIUUsgo1FrlaIzS3Se9YC62wAIKhawb211nFHKlEdgebGFLBAbPfFXiErlA7QdAQaDlRelial1UXO9m7x9A4anm0jDi-V3worGMzDoMrTaGaOSCke5oLC9LYosKGQkMZC0F1dhDurF8jcDClp4hPIA1UmmBGsUAKmtPdHgtpFAb1H46LHDzcvd_Fl4NMdPvmBQqviUVVkrvSGNQMBIH5L64PVsgGhQIx484_H3EfuT58eRDjsYzRiMYY89PJmeMxzPGIxgTz_0Q0zIm4xmTEYypd9F1OkG4RFMrafBorU47B926oVhjv5tGNTrHR61yF9MfF07nLgQa27-G_WEu3ZMD_CoOfxRfD8TBsTj6S7Kr4lrBamfREIPSMhJ0wxzaxpexTS5jm05km7X_PgMAAP__qSEMvQ==
+https://cockroachdb.github.io/distsqlplan/decode.html#eJy0k09r3DAQxe_9FGJOLchY_teDTttcSig0JS3tofigWEMQsSWjkZsswd-9WA5svd1tHfAeNeM37_cGzzNYp_Gz6pBA_oQMOOTAoQAOJXCooObQe9cgkfPTJ7PgWj-BFByM7YcwlWsOjfMI8hmCCS2ChG_qrsVbVBp9KoCDxqBMG216bzrl97uHX8Dha68sSZakk_HNECTbTRjePRLzqLRk05OCalsWTIeSCYJ65OCGcHCnoO4RZDbyM4QHsME6r9GjXkDV44kMP4zV7hF9Wi0DfPj-8e0ue7fErP7G5NCpJ9Zh5_yeDYRaskKwT-bqpaMNPbzUBbs6mypfpMrW7z37_97TPEmLzTefr2fMVzAWSVpuzlisZyxWMJZJ_Eu2ZSzXM5YrGKvkojd2gvAWqXeW8OjWTk8W0w2ivsf5YMkNvsEv3jXRZn7eRF0saKQwd7P5cW1jKwL-Kc7-KX6_EItjcf4a5xglpoK7fUBihDZIJqZlLscWlxlbXmZstdHYenzzOwAA__8vPxL7
 
 # Very simple query to make it easier to spot regressions when rewriting results
 # in test files.
